% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lads_f}
\alias{lads_f}
\title{Polygons for the Doncaster LAD outline}
\format{A dataframe with 7348 rows and 9 variables:
\describe{
 \item{long}{Longitude (east-west) coordinate of node}
 \item{lat}{Latitude (north-south) coordinate of node}
 \item{order}{plot order of node in group}
 \item{hole}{Logical: is the polygon empty (TRUE) or filled (FALSE)?}
 \item{piece}{}
 \item{id}{Identifier for polygon}
 \item{group}{Identifies which group the node belongs to}
 \item{code}{Unique code for Doncaster LAD}
 \item{name}{LAD name}
}}
\source{
See data-raw/
}
\usage{
lads_f
}
\description{
A fortified dataset containing the outline of the Doncaster LAD
}
\keyword{datasets}

