% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-management.R
\name{setup_dir}
\alias{setup_dir}
\title{Directory setup helper functions}
\usage{
setup_dir(directory)
}
\arguments{
\item{directory}{Specify a directory to test and create if it doesn't already exist}
}
\value{
Creates a directory
}
\description{
Directory setup helper functions
}
\examples{
# creates the data directory in the project's root directory
setup_dirs("data"/)
}

